Don't expose some things that should be non-public; reset device index to 0.
