+++ c/src/emokit.c
@@ -21,6 +21,19 @@
 #include <stdlib.h>
 #include <string.h>
 
+#define EMOKIT_SERIALSIZE 16
+
+#define EMOKIT_KEYSIZE 16 /* 128 bits == 16 bytes */
+
+#define EMOKIT_CONSUMER 0
+#define EMOKIT_RESEARCH 1
+
+/* ID of the feature report we need to identify the device
+   as consumer/research */
+#define EMOKIT_REPORT_ID 0
+#define EMOKIT_REPORT_SIZE 9
+
+
 const unsigned char F3_MASK[14] = {10, 11, 12, 13, 14, 15, 0, 1, 2, 3, 4, 5, 6, 7}; 
 const unsigned char FC6_MASK[14] = {214, 215, 200, 201, 202, 203, 204, 205, 206, 207, 192, 193, 194, 195};
 const unsigned char P7_MASK[14] = {84, 85, 86, 87, 72, 73, 74, 75, 76, 77, 78, 79, 64, 65};
@@ -35,6 +48,20 @@ const unsigned char AF3_MASK[14] = {46, 47, 32, 33, 34, 35, 36, 37, 38, 39, 24, 
 const unsigned char O2_MASK[14] = {140, 141, 142, 143, 128, 129, 130, 131, 132, 133, 134, 135, 120, 121};
 const unsigned char O1_MASK[14] = {102, 103, 88, 89, 90, 91, 92, 93, 94, 95, 80, 81, 82, 83};
 const unsigned char FC5_MASK[14] = {28, 29, 30, 31, 16, 17, 18, 19, 20, 21, 22, 23, 8, 9};
+
+struct emokit_device {
+	hid_device* _dev;
+	unsigned char serial[16]; // USB Dongle serial number
+	int _is_open; // Is device currently open
+	int _is_inited; // Is device current initialized
+	MCRYPT td; // mcrypt context
+	unsigned char key[EMOKIT_KEYSIZE]; // crypt key for device
+	unsigned char *block_buffer; // temporary storage for decrypt
+	int blocksize; // Size of current block
+	struct emokit_frame current_frame; // Last information received from headset
+	unsigned char raw_frame[32]; // Raw encrypted data received from headset
+	unsigned char raw_unenc_frame[32]; // Raw unencrypted data received from headset
+};
 
 struct emokit_device* emokit_create()
 {
